BORN TO UWU~~ FORCED TO OWO!!
Scorched Earth Policy once said :

I am sometimes fascinated by slaughterhouses and food processing machines. Of course they are an obvious horror, but they are also emblematic of the carnage of efficiency itself. 

Watching an eel get insta-processed into a sheet of bare viable flesh or a pile of male baby chicks evaporated into mist by an industrial culling blender within the blink of an eye summons more discomfort in the viewer than watching these creatures get eaten or starved by the cruelty of wilderness, despite these latter fates being much more painful and lengthy than the former.

What makes these machines so discomforting is in their precision, speed, and artificiality. They do not merely bring death quickly, they reduce life into geometry. They systemize the subject into substance, pure raw material to be distributed and processed accordingly in an endless conveyor belt of logistics. 

Supply and demand is a great thresher that reduces life into compartmentalized speciality. Individuality is a byproduct of generalization, a necessity when occupying the status of an apex species. It trembles at the possibility of usurpation. The moment you witness something greater than yourself, it presents a future of specialization. Existence reduced to singular purpose, a cog slotted into its perfect place to spin forever. 

What makes industrial slaughter so particularly horrifying, besides of course natural empathy for living creatures, is the mirror of possibility. If I could do this to something, something could do it to me. And a little more of something is done to us every single day. Every moment of logical decision making towards optimal outcomes pushes humanity closer towards a destiny of being processed into liquid. The delusion of reciprocity is a desperate hope, that a chicken would spare a kernel is a promise that the fox will appreciate it and find lunch elsewhere. No, whatever hunts us will not judge us by what we’ve done. It processes our sentience in the way we look at barnacles or algae. The only salvation of what you would call humanity comes through the mercy of chaos, and with it, wanton disruption. 

A great deal of this discomfort is merely a perspective of timescale. The Entropy of Efficiency threatens to reduce the scope of a lifetime into instances. What human beings could become, born, living, dying in fractions of what a human lifespan once was. Optimized, specialized, assigned, assimilated, retired, recycled. 

To the eyes of God, the light of life is a seizure inducing strobe blinking throughout eternity. In a natural ecosystem spanning across time and beyond physical space, the lifespan of the human species can be processed as an individual creature, predator to some, prey to others. The dangers lurking in the dark foliage present pieces of themselves as bait, like a lantern fish dangling its tale in the abyss. You never fully see its face until it bares its teeth for a killing blow. You only experience traces of its camouflaged existence rippling into the fabric of perception. The noises and smells alerting you to its existence come from inside you.

Do you think white blood cells understand the actions you take which introduce the diseases they fight? Do you think they experience their day of existence as an 80 year lifetime? Does a single neuron feel the sliver of influence it has to steer you in a particular direction as it experiences a fractional frame of your existence in slow motion? Have you ever purchased an article of clothing or listened to a song and realized it may be responsible for securing or wiping out the continuation of your bloodline? 

I will walk through the great machine that will slice me open, clean out my organs, crush my bones into powder, and skin me into a perfect pile of cubes. I will come out the other side in tact, unmarred, unchanged, and unbroken.